refactor array and tuple encoding;
move Tuple/Array-specific methods out of ABIType;
